# Conflicts:
#	MeteoRead/Classes/asada/Rocket.cpp
#	MeteoRead/Classes/asada/Rocket.h

+++ MeteoRead/Classes/asada/Rocket.cpp
@@ -65,5 +65,5 @@ Rect Rocket::getRocketRect(){
 void Rocket::setPower(int Pw){
 	_power = (Power)Pw;
 }
-float Rocket::getPower(){ return (int)_power; }
+int Rocket::getPower(){ return (int)_power; }
